requirements.txt
streamlit
python-dotenv
transformers
sentencepiece
torch
